# frontend/lib/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InterviewQuestion": {
        "properties": {
          "category": {
            "enum": [
              "technical",
              "general",
              "personal",
              "project"
            ],
            "title": "Category",
            "type": "string"
          },
          "dimension": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dimension"
          },
          "source": {
            "$ref": "#/components/schemas/QuestionSource"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "topic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic"
          }
        },
        "required": [
          "text",
          "category",
          "source"
        ],
        "title": "InterviewQuestion",
        "type": "object"
      },
      "InterviewSet": {
        "properties": {
          "general": {
            "items": {
              "$ref": "#/components/schemas/InterviewQuestion"
            },
            "title": "General",
            "type": "array"
          },
          "personal": {
            "items": {
              "$ref": "#/components/schemas/InterviewQuestion"
            },
            "title": "Personal",
            "type": "array"
          },
          "projects": {
            "items": {
              "$ref": "#/components/schemas/ProjectQuestions"
            },
            "title": "Projects",
            "type": "array"
          },
          "technical": {
            "items": {
              "$ref": "#/components/schemas/InterviewQuestion"
            },
            "title": "Technical",
            "type": "array"
          }
        },
        "required": [
          "personal",
          "projects",
          "technical",
          "general"
        ],
        "title": "InterviewSet",
        "type": "object"
      },
...
      "ProjectQuestions": {
        "properties": {
          "project": {
            "title": "Project",
            "type": "string"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/InterviewQuestion"
            },
            "title": "Questions",
            "type": "array"
          },
          "technologies": {
            "items": {
              "type": "string"
            },
            "title": "Technologies",
            "type": "array"
          }
        },
        "required": [
          "project",
          "technologies",
          "questions"
        ],
        "title": "ProjectQuestions",
        "type": "object"
      },
      "QuestionSource": {
        "properties": {
          "kind": {
            "enum": [
              "github",
              "template"
            ],
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "license": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License"
          },
          "template_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Id"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "required": [
          "kind",
          "label",
          "url",
          "license"
        ],
        "title": "QuestionSource",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Resume Optimizer API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/analyses/{analysis_id}/interview": {
      "get": {
        "operationId": "get_interview_analyses__analysis_id__interview_get",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Analysis Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewSet"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Interview",
        "tags": [
          "interview"
        ]
      },
      "post": {
        "description": "\"Prepare Me for Interview\". Idempotent: returns the stored set if there is one.",
        "operationId": "prepare_interview_analyses__analysis_id__interview_post",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Analysis Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewSet"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Prepare Interview",
        "tags": [
          "interview"
        ]
      }
    },
...
  }
}